refactor(OnlineGiftPackage): use `ConsoleInfo`  to replace `WriteLine`, remove namespace `在线礼包`
