+++ src/OnlineGiftPackage/Config.cs
@@ -1,6 +1,5 @@
 ﻿using Newtonsoft.Json;
 using TShockAPI;
-using 在线礼包;
 
 namespace OnlineGiftPackage;
